requirements.txt
absl-py==2.1.0
accelerate==0.20.3
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
anykeystore==0.2
apex==0.9.10.dev0
async-timeout==4.0.2
attrs==23.1.0
base58==2.1.1
blinker==1.6.3
blis==0.7.11
boto3==1.28.62
botocore==1.31.62
cached-path==1.2.0
cached-property==1.5.2
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
cmake==3.26.4
codebleu
colorama==0.4.6
commonmark==0.9.1
contourpy==1.1.1
cryptacular==1.6.2
cycler==0.12.1
cymem==2.0.8
datasets==2.13.0
defusedxml==0.7.1
dill==0.3.6
docker-pycreds==0.4.0
docopt==0.6.2
edit-distance==1.0.6
./en_core_web_sm-3.4.0-py3-none-any.whl
evaluate==0.4.0
exceptiongroup==1.1.3
fairscale==0.4.6
faiss-cpu==1.7.4
faiss-gpu==1.7.2
filelock==3.7.1
fitlog==0.9.15
Flask==3.0.0
fonttools==4.43.1
frozenlist==1.3.3
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.37
google-api-core==2.12.0
google-auth==2.23.3
google-cloud-core==2.3.3
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.0
h11==0.12.0
h5py==3.10.0
httpcore==0.15.0
httpx==0.23.0
huggingface-hub
hupper==1.12
hydra-core==1.2.0
idna==3.4
inflect==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
javalang==0.13.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonlines==3.1.0
jsonnet==0.20.0
kiwisolver==1.4.5
langcodes==3.3.0
lit==16.0.6
lmdb==1.4.1
MarkupSafe==2.1.3
matplotlib==3.6.0
matplotlib-inline==0.1.6
more-itertools==10.1.0
mosestokenizer==1.0.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.10
networkx==3.1
nltk==3.8.1
numpy==1.23.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
openfile==0.0.7
overrides==6.2.0
packaging==23.1
pandas==2.0.2
PasteDeploy==3.0.1
pathtools==0.1.2
pathy==0.10.2
pbkdf2==1.3
Pillow==10.0.1
plaster==1.1.2
plaster-pastedeploy==1.0.1
pluggy==1.3.0
preshed==3.0.9
prettytable==3.9.0
progressbar==2.5
promise==2.3
protobuf==3.20.3
psutil==5.9.5
pyarrow==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.8.2
pyext==0.7
Pygments==2.16.1
pyparsing==3.1.1
pypng==0.20220715.0
pyramid==2.0.2
pyramid-mailer==0.15.1
PySocks==1.7.1
pytest==7.4.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
regex==2023.6.3
repoze.sendmail==4.4.1
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rfc3986==1.5.0
rich==12.6.0
rouge-score==0.1.2
rsa==4.9
s3transfer==0.7.0
sacremoses==0.0.53
safetensors==0.4.3
scikit-learn==1.3.1
scipy==1.11.3
sentencepiece==0.1.99
sentry-sdk==1.32.0
setproctitle==1.3.3
shadowsocksr-cli==2.1.10
shortuuid==1.0.11
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
spacy==3.4.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.21
srsly==2.4.8
sympy==1.12
tensorboardX==2.6.2.2
termcolor==1.1.0
thinc==8.1.0
threadpoolctl==3.2.0
tokenizers==0.12.1
tomli==2.0.1
toolwrapper==2.1.0
tqdm==4.65.0
traitlets==5.11.2
transaction==3.1.0
translationstring==1.4
tree-sitter==0.21.3
triton==2.0.0
typer==0.4.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.3
velruse==1.1.1
venusian==3.0.0
wandb==0.12.21
wasabi==0.10.1
wcwidth==0.2.6
WebOb==1.8.7
Werkzeug==3.0.1
wget==3.2
WTForms==3.1.0
wtforms-recaptcha==0.3.2
xxhash==3.2.0
yarl==1.9.2
zope.deprecation==5.0
zope.interface==6.1
zope.sqlalchemy==3.1

